betting_app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
playwright
beautifulsoup4
requests
pandas
numpy
scikit-learn
vaderSentiment
python-dotenv
python-dateutil
jinja2
feedparser
